refactor(exception): reorganize exception classes and update package structure
