+++ src/main/java/com/pcabp/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package com.pcabp.core.common.exception;
+package com.pcabp.exception;
 
 public class BusinessException extends RestException {
