+++ src/main/java/com/pcabp/exception/SystemException.java
@@ -1,4 +1,4 @@
-package com.pcabp.core.common.exception;
+package com.pcabp.exception;
 
 public class SystemException extends RestException {
