+++ src/main/java/com/pcabp/model/RestResponse.java
@@ -1,4 +1,4 @@
-package com.pcabp.core.common.dto;
+package com.pcabp.model;
 
 import lombok.*;
