+++ src/main/java/com/pcabp/model/RestResponseType.java
@@ -1,4 +1,4 @@
-package com.pcabp.core.common.dto;
+package com.pcabp.model;
 
 public enum RestResponseType {
     SUCCESS,
